Initial Commit
Embarassingly, This is the first commit despite having done numerous
rewrites of the original code base. The original code base was deleted
in a fit of rage and therefore this serves as the only documented
version of the Crowfall Community Bot. This version implements
modularized commands, powers library, and administrative permissions
commands. Fully one third of planned features are now working to a
degree.